chore: Add a to-do item to use events for *_checked APIs
This would have two advantages:

1. The `*_checked` APIs would have the same consistency behaviour as
   CloudFormation itself, and we could eliminate the `Conflict` error
   case.
2. It should be possible to share the bulk of the logic across all
   stack-based `*_checked` and `*_stream` functions.

The 'disadvantage' would be that those APIs would return `StackEvent`s
rather than `Stack`s. Since much of the salient information is present
in `StackEvent`, including the `stack_id` which can be used to reliably
obtain further information, this is probably still an overall win.