refactor!: Simplify stream error types
This is a step towards making the `raw` API 'more raw'. The need for
custom errors types has been reduced by returning a `Future` from the
`*_stream` methods. The output of the future captures the
'initialisation' errors so that the stream only has to deal with the
`DescribeStackEvents` errors. This has little impact on the API's
usability (what's an extra `.await`?) and makes error handling more
logical (initialisation errors aren't possible after initialisation).

Notably, no errors will be emitted if the overall stack operation fails
(e.g. it settles with a failed status). This feels like a higher-level
feature that should be implemented in another part of the library.

BREAKING CHANGE: `CreateStackStreamError` and `UpdateStackStreamError`
have been removed. `DeleteStackStreamError` and
`ExecuteChangeSetStreamError` have had their `Failed`, `Conflict`, and
`DescribeStackEvents` variants removed. The `*_stream` methods now
return a `Future`, and the streams now use
`RusotoError<DescribeStackEventsError>` for errors.