feat: Add some status types and helpers
The lib now exposes enums for each of change set, stack, and resource
statuses. The enums all implement `Display` and `FromStr` to convert to
and from the strings returned by the CloudFormation API.

They additionally implement a `Status` trait which provides methods for
testing whether a status is terminal and what the 'sentiment' of the
status is (positive, neutral, or negative). These can be used to
determine whether a change set/stack/resource has settled, and if so
whether the operation succeeded. They would also be useful for
colorising statuses like the AWS Console does.