feat: Add a struct representing a stack event
The `StackEvent` type presents all the information from the
`rusoto_cloudformation::StackEvent` type, but with more accurate
constraints and richer types. The main difference is that the top-level
representation is as an enum with `Stack` and `Resource` variants. This
is necessary because we have different types for `StackStatus` and
`ResourceStatus`, which itself reflects CloudFormation's own model.

To make using the enum easier, all the properties from
`StackEventDetails` have accessors on the `StackEvent` enum. There's
additionally an accessor for the `resource_status` which returns a
`Status` trait object. Finally there's an `is_terminal` method that
returns true for stack events with a terminal status, indicating that
the relevant stack operation has finished.