feat: Add scaffolding for `apply` operation
There is no implementation yet, but the input and output types have been
defined. The `ApplyInput` type is a common subset of the
`CreateStackInput` and `UpdateStackInput` structs, with tighter
constraints and richer types where necessary. The `ApplyOutput` is a
simplified version of `Stack` focused on the state of the stack after
the operation.