feat: Add more scaffolding for `apply`
This is sketching out the `apply` API a bit more, including defining the
stream item and error type.

The current thinking is that the returned `Apply` struct will implement
both:

- `Future<Output = Result<ApplyOutput, ApplyError>>`
- `Stream<Item = Result<ApplyEvent, ApplyError>>`

The `ApplyEvent` type is an enum that can contain either a `StackEvent`
or `ApplyOutput`. The principle would be that the `Stream` will emit
`StackEvent`s until it has reached a terminal event, when it will yield
the `ApplyOutput`. It remains to be seen how sensible this is... It
doesn't feel great that the type system can't capture the fact that
`ApplyOutput` must be the final item in the stream.