feat: Implement `Future` for `Apply`
This is the first implementation of the `apply` high-level operation
that performs the following actions:

1. Create a change set (support idempotency by retrying `CREATE`s as
   `UPDATE` if the stack already exists).
2. Wait for the change set to settle (support idempotency by handling
   failures due to empty change sets).
3. Execute the change set (if it wasn't empty), and wait for the stack
   to settle.
4. Return the new stack state, or a detailed error if anything went
   wrong.

The operation is exposed via a new top-level `CloudFormationExt` trait,
which has an `apply` method that calls `Apply::new`.

The commit got a bit out of hand, there were a lot of additional
utilities and trait implementations needed in `event` and `status` as
well as `apply` itself.