refactor!: Remove `raw` module
This was always intended to be wrapped by higher-level APIs. Initially
it was thought that exposing the `raw` publicly could be useful to users
in the event that the higher-level APIs were not quite applicable for
their use-case. As the library has developed, however, this has felt
like a bit of a maintenance risk, since the line between what is 'raw'
and what is 'high-level' is a bit unclear, and might change if
additional APIs are added to the crate.

A better way of letting users 'escape' from the high level APIs would be
to expose (more of) the methods for converting from rusoto types to
library types, such as `StackEvent::from_raw`, but it's not clear if
this would really be useful yet and these could be exposes
backwards-compatibly in future.

BREAKING CHANGE: The entire `raw` module has been removed.