refactor: Move apply event handling into a dedicated Stream
The `StackEvent` handling in `Apply::new` was quite verbose, which made
it hard to grok the control flow. In an attempt to makes things easier,
an `ExecuteChangeSet` struct has been created which wraps an
`execute_change_set` stream, along with all the stack event error
recording. This means that `Apply::new` can now simply pull values from
the `ExecuteChangeSet` stream, and once it's exhaused it attempts to
extract an `ApplyOutput`, which may return an `Err` depending on what
errors were recorded.