chore!: Rename the library to cloudformatious
It doesn't seem like a good idea to use `rusoto` in the name, since this
is in no way affiliated with that crate or its authors.

`cloudformatious` is a made up word – it even has very few hits on
Google! – but it does kind of convey 'of or pertaining to
CloudFormation'.