refactor!: Remove ApplyEvent
This removes the `ApplyEvent` enum and updates the `Stream` impl on
`Apply` to simply yield `StackEvent`s instead. Output and errors now
need to be obtained via the `Future` impl. This does mean making sure
not to consume the `Apply` struct when streaming (particularly with
certain combinators, as can be seen in the tests).

BREAKING CHANGE: `ApplyEvent` has been removed. `Apply` now implements
`Stream<Item = StackEvent>`.