refactor: Move stack event streaming into stack module
This makes the `stack` module responsible for streaming stack events and
aggregating error/warning information. This puts all the stack
error/warning handling inside the same module, which is nice, and takes
quite a lot of weight out of the `apply_stack` and `change_set` modules
which no longer need to implement error/warning handling or stack event
streaming, respectively.

As an added bonus, this is all backwards compatible 🎉