refactor: Remove special handling of client_request_token
Originally it was hoped that we could use this rather than timestamps to
correlate events with operations. Since we would anyway have to use this
in conjunction with a timestamp in order to filter out events that have
already been processed, this isn't currently a high priority. So we may
as well remove this handling until we actually want to use it.