feat: Flesh out change_set and make it public
The `change_set` module is now public, and features a more complete
`ChangeSet` struct and a smorgasbord of supporting structs and enums.

There are not yet any APIs for constructing these types. It's
anticipated that there will be some mechanism for inspecting
`ChangeSet`s prior to executing `ApplyStack` (and maybe even
`DeleteStack`) operations.