feat: Add an API to get the ChangeSet that will be applied
This allows consumers to interrogate the change set before continuing to
execute it. The API is a bit rough, in that there is an implicit order
in which operations must be used to avoid panics (`change_set()` ->
`events()` -> `.await`). Perhaps the wrapping types could consume the
previous ones... but this would be a breaking change and will be
considered separately.