feat!: Implement Copy for change_set::Action
This changes how the `scope` field of the `change_set::Action::Modify`
enum variant is represented, from `HashSet` to `EnumSet`, in order to
allow `change_set::Action` to implement `Copy`. This is desirable as
it's meant to be a simple marker that can easily be passed around,
compared, etc.

BREAKING CHANGE: The type of the `scope` field of the
`change_set::Action::Modify` enum variant has changed from `HashSet` to
`EnumSet`. Since `EnumSet` implements `FromIterator`, `IntoIterator`,
and many other set methods this should be transparent in many cases, but
it's still a change to the public API.