refactor!: Move ResourceChange::details into Action::Modify
It was intended that all the modify-only `ResourceChange` properties
should live on the `Action::Modify` enum variant. `details` was missed,
but it is documented as only being relevant for modify actions.

Rather than adding additional properties directly to `Action::Modify`,
all of its properties have been moved to a `ModifyDetail` struct, and
`Action::Modify` itself is now a tuple-variant with a single field.

Sadly, this undoes apart of d2f5268 since it no longer makes sense for
`Action::Modify` to be `Copy` (though `ModifyDetail`'s `scope` field
remains an `EnumSet`).

BREAKING CHANGE: `Action::Modify` is now a tuple-variant with a single
`ModifyDetail` field, which contains the variants previous fields. The
`details` field of `ResourceChange` has also moved to the `ModifyDetail`
struct.