test: Add a test to check change set creation is idempotent
This was inspired by paranoia from a previous implementation that was
based on trying to describe the stack to determine its existence in
order to decide whether to create or update. That approach requires
explicit handling of the `REVIEW_IN_PROGRESS` state. Since we have no
explicit handling for that state, I wondered if it would work.

Turns out it does, since we're using 'create or update' based on whether
the create fails with an "already exists" error.