feat: Add method to StackFailure to get StatusReason
This adds a `StackFailure::stack_status_reason` method, which returns a
`StatusReason` wrapping the underlying `stack_status_reason` string.
This provides an analogue to `StackEventDetails::resource_status_reason`
for the stack failure itself (without this there's no other way for
users to create a `StatusReason` for the stack failure, since all its
constructors are crate-private).