fix!: Fix handling of tag changes in change sets
We had assumed that `change_source` could never be null in
`ResourceChangeDetail`. Sadly, it *is* null when performing an update
that only affects a resource's tags via the change set's input (as in
the new test). As such, `ResourceChangeDetail::change_source` has been
made an `Option`.

Fixing this immediately revealed another incorrect assumption: that
`ResourceTargetDefinition::requires_recreation` would always be null for
non-`Properties` target attributes. Again, this is not the case when
performing an update that only affects a resource's tags via the change
set's input. Rather than changing the API in this case, the assertion
around `requires_recreation` has been changed to assert either `None` or
`Some("Never")` – so we still assume that only `Properties` changes can
cause recreation, and keep it out of the API surface in other cases.

BREAKING CHANGE: The type of
`change_set::ResourceChangeDetail::change_source` has been changed from
`ChangeSource` to `Option<ChangeSource>`.