fix: simplify the principal-less missing permission regex
There is at least one other possible "trailer" for the error message,
which is "Access Denied" from attempting to create an S3 Bucket without
permission. Rather than continuing to play whac-a-mole we now rely on
just the `API: ` prefix followed by something that looks like a
permission.

This feels a bit flaky, but without a standard for permissions errors
it's always going to be a bit best-effort.