feat: handle encoded authorization failure messages
This adds a new `status_reason::EncodedAuthorizationMessage` type, which
wraps an encoded authorization message string. The message can be
decoded using `decode` by passing in an STS client.

The `status_reason::MissingPermission` now has an optional
`encoded_authorization_message` field, reflecting the fact that
permission errors sometimes have an encoded authorization message
embedded in them as well.

There's also an additional `StatusReasonDetail::AuthorizationFailure`
variant that includes just an `AuthorizationFailureMessage`. This is
returned for status reasons that include an authorization failure
message but no unencoded permission info. This is not a breaking change
since the enum is non-exhaustive.

The upshot of this is more useful information can be obtained through
the API for authorization errors, which should allow for more useful
diagnostics!