refactor!: switch from rusoto to the official aws-sdk
This was a fairly mechanical switch overall (e.g. from `Input` structs
to builder APIs), but changes to the 'entrypoint' are more significant.

Specifically the aws-sdk does not use a trait for behaviour, and instead
uses inherent impls on `Client` structs. This meant that using a trait
for the cloudformatious API became significantly more work since it
would also need to cover the lower-level APIs needed to drive
`apply_stack` and `delete_stack`. Furthermore, the trait-based approach
doesn't really offer anything when the trait is "sealed", and just makes
calling more awkward.

Follow-up work should try to reduce the surface area of aws-sdk that is
exposed through the API, and potentially align conventions with the
aws-sdk in terms of using builders instead of inputs.

BREAKING CHANGE: The `CloudFormatious` trait has been removed, so a
`rusoto_cloudformation::CloudFormationClient` can no longer be used to
call cloudformatious APIs. Instead, you should construct a
`cloudformatious::Client` from an `aws_config::SdkConfig`.
`EncodedAuthorizationMessage::decode` now needs to be passed an instance
of `aws_sdk_sts::Client`.