feat!: add `ApplyStackError::Blocked` variant
Attempting to perform an `apply_stack` operation on a stack in a blocked
status will now cause an `ApplyStackError::Blocked { .. }` error to be
returned. This offers direct access to the `BlockedStackStatus`, making
it easier to implement appropriate handling, if desired.

BREAKING CHANGE: A `Blocked` variant has been added to the
`ApplyStackError` enum, which is exhaustive.