feat: add `Client::resume_apply_stack`
This essentially runs the `apply_stack` logic, except that the change set
creation is skipped.

One curiosity is that this could be ran against a stack that settled some
time in the past and it would still replay the events from the operation.
This suggests the naming might not be quite right, but the functionality
could be useful.